Use nacl.secretbox for cookies encryption
AES-GCM only has 12 bytes nonces. For completely random nonces this is a
little low, thus change encryption to nacl with Curve25519, XSalsa20 and
Poly1305 to encrypt and authenticate the cookie values in the access
token (this mode uses 24 byte nonces).

Also get rid of the seperate nonce field. Instead the nonce is prepended
to the cipher text.