Implement identifier and kc backend
This commit adds a native login area to Konnect which support multiple
backends. A backend implementation which directly connects to Kopano
Core server via SOAP is added.

The login area is implemented as identifier together with a web
application and the required API endpoints to support that web
application. The web app is implemented with React and uses Material UI
for the HTML representation, providing a fully functional solution to
sign into Kopano together with a OIDC flow to retrieve ID token and/or
access token to consume further APIs.