identifier: Implement uuid attribute support
To allow mapping between different services connected to the same LDAP
server as Konnect, it is sometimes useful to have an additional unique
user ID claim.

This change addes the `konnect/uuid` scope which results in the
`kc.uuid` claim which contains a unique user ID when supported by the
identifier backend.

Unique user ID support is added to the LDAP identifier backend. For
configuration, the environment variables LDAP_UUID_ATTRIBUTE and
LDAP_UUID_ATTRIBUTE_TYPE are now supported. If the attribute type is set
to `binary` the value will be returned base64 encoded.