identifier: Add UUID decode support to ldap uuid
With active directory, the unique user ID is often used from the
objectGUID attribute which is a UUID version 4. Instead of simply base64
encoding its binary value, a new attribute type `uuid` is now available
which can parese UUID version 1, 2, 3, 4 and 5 and return the proper
string encoding for those values.