identifier: Add Client registry and validation
This change adds a YAML configuration which can be used to provide
information for know OIDC clients. The hardcoded client_ids have been
removed.

To retain the previous behavior, take a look at the
`identifier-regisration.yaml.in` file and provide it via the new
`--dentifier-registration-conf` parameter. Konnect will load and
register all clients defined in that file on startup.

For now, Konnect needs to restart when the YAML file is changed;
automatic reload or hot reload via signal will be implemented with a
later commit.