identifier: Use no_uid_auth flag for logon to kc
When using the UNIX socket to connect to Kopano server and the
process is running as the same unix user, special privileges
are obtained by the client, so user and password are not checked. This
change enforces checking of passwords in all cases even if the process
is running as the same unix user like the configured Kopano server
backend.