docker: Allow Docker to switch user at runtime
To be able to read secrets from arbitrary locations while at the same
time avoid running as root, this change adds su-exec and the environment
variables KONNECTD_USER and KONNECTD_GROUP to allow switching the
runtime user and group of konnectd just before startup but after reading
secrets. Default behavior is unchanged, running as nobody/nogroup and it
is also still possible to run directly as an unpriviledged user.