Add checks and consent to end session support
Ending a session requires various checks of the request to avoid
allowing everyone to sign out. This change adds checks for the redirect
uri, client ID and also adds a distinction for non=trusted clients which
now not directly sign-out but instead show a consent message to the user
which requests confirmation.