oidc: Add support for url-safe sub via scope
To allow integration with third party applications easily it is often
required that the Subject in the ID Token and Userinfo are URL safe
values. This changes adds the new scope `konnect/hashed_sub` which
enables this behavior for ID Tokens and Userinfo reponses. Access tokens
are Konnect specific and never used by third party code and thus retain
the existing behavior where the sub claim is the raw backend user
Subject.

To generate a public Subject consisently, Blacke2B 512 with static MAC
for easy validation. Also the public subjects have a suffix to identify
them as @konnect for potential future use.