oidc: Add support for additional userinfo claims
The userinfo endpoint is a flexible one that can return additional
claims when requested by a corresponding scope or claims request. This
change makes the userinfo endpoint result flexible and introduces the
possibility for backends to return additional claims based on incoming
authorized scopes.

The first additional claim returned is the Kopano Groupware Core backend
id when using the kc identifier backend and the kopano/gc scope is
requested.