oidc: Implement browser state and session state
This implements the session state value and endpoints as specified in
the OpenID Connect Session Management 1.0 Draft 28
(https://openid.net/specs/openid-connect-session-1_0.html). Browser
state is stored in as a Blake2b-256 hash cookie value which can be read
by the OP IFrame implementation.

The client side session state uses SHA256 which is easily available in
Javascript and other clients.

Since becase of the same-origin policy, validation cross domain origins
is impossible in modern browsers, the check session iframe only reacts
on messages if the source or the source parent is its equal to its own
parent window.