oidc: Implement `sid` ID Token claim
The `sid` claim in ID Tokens define a unique session ID for an OIDC
client at the OP for the RP. This change adds a cookie at the OP to
store an encrypted session ID which is can be used by RP clients to
identify if something has changed. While we currently do not use this
value for anything, some clients (namely oidc-client-js) wrongfull
depend on the `sid` claim when checking the session state (see
https://github.com/IdentityModel/oidc-client-js/issues/654 for
reference). With this change, Kopano Konnect is compatible with the
SessionMonitor logic in oidc-client-js including `session_state` and
`sid` change detection.