oidc: Make subject URL safe by default
For compatibility reasons with clients, it is required that the Subject
of the ID token is URL safe. This is not possible with the plain
EntryIDs used by Kopano Groupware. This change changes the default that
the Subject of tokens is a url safe hased variant of the underlaying
user provides real identifying subject.

The previous scope `konnect/hash_sub` has been removed since it is now
the default (adding it has no effect). To retain the old behavior if
required a client can request the new scope `konnect/raw_sub`.

This is a breaking change since all subjects change. If the subject is
used as auth identifier for users, all users will be new users.