identifier, oidc: Add support for backend identity provider sessions
A backend server might have sessions which grant access to user
information. This change adds a backend sesseion reference to identifier
logon cookies and to the oidc access token, allowing to use the same
session in the user info endpoint and also for Kopano clients which know
how to get that value from access token.