identifier: Add configuration for kc session timeout
When using per user sessions together with the kc identifier backend,
the access tokens and sign in is essentially bound to the kc server
session which auto expires on no activity. Thus a new environment
variable KOPANO_SERVER_SESSION_TIMEOUT is added which allows to specify
the session timeout in seconds (defaults to 300). As a consequence the
validity of access tokens cannot be longer than the session timeout
which neatly enforces clients to refresh the underlaying session when
the access token needs refreshing. The startup logic limits the access
token duration automatically if required.