oidc, identifier: Use kcoidc auth to kc for kc sessions
To avoid signout of all users when kc server is restarted this change
adds support for Konnect to establish kcoidc sessions with kc backend
internally. Whenever required, Konnect creates an approriate access
token and does a new ssoLogon at the kc backend for the given user.

This should become the recommended mode for Konnect in production, to
avoid running Konnect as a Kopano admin (which is still also supported).