oidc: Use correct Salt length with RSA-PSS signatures
RSA-PSS signatures require a salt length. Unfortunately RFC 3447 does
not specify the salt length and thus people might use different salt
lengths resulting in invalid signatures. This change sets the salt
length to be the same as the hash size following the recommendation in
https://www.ietf.org/mail-archive/web/jose/current/msg02901.html (It
was pointed out to me that we failed to specify the salt size for RSA
PSS signatures.  RFC 3447 says that “Typical salt lengths in octets are
hLen (the length of the output of the hash function Hash) and 0.”
Having looked into it a bit I believe that it would be appropriate for
us to specify that the salt length be the same as the output size of the
hash function used. So 256 for PS256, 384 for PS384, and 512 for PS512.)