scripts: Ensure correct owner when creating paths
Konnect is running as a restricted user, thus it is required that this
restricted user can actually access the keys from the validation keys
path. When setup is used to create this path, the owner ship of that
folder is now set which lets konnect later even open the keys from
there.