Implement support for request objects with OIDC authorize
To support additional signed claims, this commit is the first of a
series supporting that. OIDC core standard supports a request object
with the authorize endpoint. That request object is a JWT which can be
optionally signed. This commit adds the request object with its standard
OIDC claims for unsigned JWTs.