Add validation for the sub requested claim
When the authorization request contains the sub claim request with a
value, the server is oblidged to reject the request if the authenticated
user is not the one requested. This change implements this behavior by
matching the subject of the authenticated user to the sub requested
claim value when it was given.