Add multi server support for kc backend
Kopano user entry IDs are ABEIDs which can be different in multiserver
setups even for the same user. This change uses on the exID part of the
ABEID of a user to generate the external facing public subject in the
OIDC ID Token. Also internally the kc backend now uses multi server
aware logic defining the user properly.

Fixes: KC-1384