Support trusted client scopes in secure oidc request
To allow additional scopes and guest access, a ODIC client can provide a
signed requres object. This request might request additional scopes and
claims which grant additional access. To limit this access grant, this
change adds a new `trusted_scope` configuration to the identifier
registration which can be used to grant guest access with additional
scopes.

To allow guest access for a specific client, it is now required to
configure that client with trusted_scopes that include
`konnect/guestok`. Registered clients without that scope and a signed
request will no longer authorize as guests.