Implement PKCE code challenges as defined in RFC 7636
This change adds support for code challenges when the OAuth2 code flow
is used. This additional security mechanism can be used to ensure that
the client completing the token request with a code is the client which
initiated the authorization. This change adds support for the code
challenge method S256 to the authorize and token requests as
optional parameters.

Reference: https://tools.ietf.org/html/rfc7636