Make LDAP subject source configurable
By default the public subject of users from the LDAP backend is based on
the DN of the corresponding users LDAP record. Since the DN might change
and certain LDAP trees might have better unchanging unique values to
base the subject on, this change adds the new environment variable
`LDAP_SUB_ATTRIBUTES`. When set, it defines what LDAP fields are used to
derive the public subject from. The value is a space separated list of
LDAP field names. For example `uid uidNumber` will use both of those
fields to derive the subject. When the environment variable is not set
or empty, the bevhavior is unchanged and the subject of LDAP users will
be derived from their DN.

Reference: https://github.com/Kopano-dev/konnect/issues/3