Add support for konnect/id scope with LDAP backends
The konnect/id scope can provide a numeric ID to users. Some LDAP
servers might have such an attribute and this change adds support for it
via an optional environment variable `LDAP_UIDNUMBER_ATTRIBUTE`. If this
env var is non-empty, it will be used as attribute name to find a
numeric value to be used as id claim. If this is enabled, all users must
return this field with a numeric value. When the variable is not set,
the konnect/id scope is not advertised and no id claim is returned for
LDAP users.

Fixes: KK-1