Integrate Delve debugger support via `make dlv`
Running the debugger becomes a lot easier with this change. It simply
adds a `make dlv` target which attaches the Devle debugger to a
currently running process of `bin/konnectd`.

So a typical development workflow would be to run `./bin/konnectd` in
one terminal with all its parameters and stuff. In another terminal run
`make dlv` to connect the debugger to it.