Force prompt=none for sencodary authorize after external authority auth
When an external authority is authorization is reqquested with prompt
parameter that said prompt parameter must only be used by the external
authority and not by the local authority callback authorization. This
change always sets the prompt to none on the secondary local
authorization callback redirect to avoid potential loops when a
interactive prompt is requested.