Strip issuer subpath for OIDC url endpoints
URL endpoints subpath is solely to be configured by the `uri-base-path`
parameter. This change strips the subpath from the configured `iss`
value to avoid duplicating a subpath value. When running the OIDC
provider in a subpath, both `iss` and `uri-base-path` must now be set in
a matching fashion so that the endpoint urls controlled by
`uri-base-path` match what is expected. Furthermore it is no longer
technically required to use a subpath in the `iss` value but for the
sake of uniqueness it is strongly recommended.