Detect browser state change issues
In some cases an existing browser state might not change which based on
the client implementation might lead to an endless 'changed' loop.

I have seen this happending when an old browser state cookie was not
replaced by the browser after the SameSite=None change. It seems that
some older browser could have bugs with SameSite=None leading to the
browser session not updatin the browser state cookie.

With this fix, the browser state frame remembers the last seen browser
state event and returns error when it sees the same thing again.