Update oidc-go to fix pkce Base64URL padding
It was discovered that oidc-client-js pkce does not work with Konnect
and Konnect always reports `invalid code challenge`. It turned out the
oidc-go library incorrectly compared Base64URL with padding while the
specification wants it without padding. This change updates oidc-go to
the 0.3.1 which includes the fix for this issue.

Reference: https://tools.ietf.org/html/rfc7636#appendix-A