Make konnect usable as library
This commit introduces several refactorings to make this project embedable as library.
To achive this the bootstrap package has been moved from cmd in to a seperate package
and all dependncies in the bootstrap process have been decoupled from the environment.

The initialization process has been abstracted in to a single Boot() method which can be called
by consuming code.

Further a interface was introduced for the bootstrap struct to expose certain extension-points
to the consumer.