Implement delegation of end session to external authority
External authority support for sign in is one thing, but in certain
cases it might be desired that a sign out in Konnect also triggers a
sign out in an external authority. This change brings this for SAML2
external authorities. Enable it by setting `end_session_enabled` in the
corresponding authority definition.