Generate random endsession state for external authority
When a client does not provide a state when calling OIDC endsession
endpoint while at the same time Konnect is configured to use an external
authority which requires a state to keep track of the request this
change makes sure that were actually is a state. This cures are
regression introduced in 0.33.4 which actually added support to pass
through the original state, but forgot about the case when there is no
such state provided.