Improve checks for Basic auth data in token requests
If a client sends the Authorization header to the token endpoint, this
usually means that it provides client_id and secret via
client_secret_basic authentication method. If a web client uses the
token endpoint, the Authorization header value might not always actually
be meant for the token endpoint and thus is rejected. Since we cannot
change what the browser / user is actually doing, this change adds rules
which allows Konnect to ignore client credentials sent to the token
endpoint via Authorization header in certain cases. This means the
provided values are only used if no client_id is passed via the request
query or if the query provided client_id actually matches the user part
of the basic auth decoded value.