Allow configuration of expiration of dynamic client_secret values
The client_secret for dynamically registered clients is generated by the
server and can have a duration. The duration was hardcoded to 1 hour
which makes the dynamic client feature not really usable for use
together with long term refresh tokens. With this change, the dynamic
client_secret values no longer expire. If expiration is desired, the new
commandline parameter `dynamic-client-secret-expiration` can be used to
specific customer behavior as desired.